requirements.txt
google-generativeai
chromadb
pypdf
langchain-google-genai
python-dotenv
streamlit
langchain_groq
langchain-community
langchain_experimental
duckduckgo-search
arxiv
xmltodict
langchain
